share studio run status UI
