+++ apps/studio/src/routes/projects/$projectId_/runs/$runId.tsx
@@ -8,6 +8,7 @@ import { useState } from 'react';
 import { ResumeRunActions } from '~/components/ResumeRunActions';
 import { RunDetail } from '~/components/RunDetail';
 import { RunEvalModal } from '~/components/RunEvalModal';
+import { RunStatusIndicator } from '~/components/RunStatusIndicator';
 import { StopRunButton } from '~/components/StopRunButton';
 import { useProjectRunDetail, useStudioConfig } from '~/lib/api';
 
@@ -92,6 +93,7 @@ function ProjectRunDetailPage() {
               runStatus={runStatus}
             />
           )}
+          {runStatus && <RunStatusIndicator status={runStatus} />}
           {!isReadOnly && !isActiveRun && (
             <button
               type="button"
